requirements.txt
scapy==2.4.5           # Для сниффинга сетевого трафика
psutil==5.9.0          # Для получения информации о соединениях
PyYAML==6.0            # Для работы с YAML-конфигами
python-dotenv==1.0.0   # Для загрузки переменных окружения
elasticsearch==8.9.0   # Если будешь отправлять напрямую в Elasticsearch
python-logstash==0.4.6      # Альтернатива для отправки в Logstash
requests==2.31.0       # Для HTTP-запросов, если нужно что-то проверять
tldextract==5.1.2      # Для извлечения домена из URL